new feature: menu cadastrar
